backend/requirements.txt
flask==3.1.1
flask-socketio==5.5.1
flask-cors==5.0.1
python-socketio==5.12.1
python-engineio==4.11.2
eventlet==0.37.0
SQLAlchemy==2.0.48
python-dotenv==1.2.2
google-genai
werkzeug==3.1.3
